third_party/blink/web_tests/flag-specific/disable-layout-ng/svg/custom/createelement-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: Cannot read properties of null (reading 'setAttribute')
